HAK-42 change lease rate to BigDecimal
-

+++ src/test/java/nl/hakktastic/leaseacarapi/testdata/LeaseRateTestData.java
@@ -4,13 +4,14 @@ import nl.hakktastic.leaseacarapi.bean.CarBean;
 import nl.hakktastic.leaseacarapi.bean.CustomerBean;
 import nl.hakktastic.leaseacarapi.bean.InterestRateBean;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
 /** Test Data for Lease Rate Service. */
 public final class LeaseRateTestData {
 
-  public static final Double LEASE_RATE_VALID = 574.38;
+  public static final BigDecimal LEASE_RATE_VALID = BigDecimal.valueOf(574.38);
   public static final int mileage = 35000;
   public static final int duration = 48;
